+++ layout/main.py
@@ -129,7 +129,7 @@ def get_graph_options(data_type, graph_type, preset_options=None):
             component = dcc.Checklist(
                 id={"name": option, "type": f"graph_{data_type}_option"},
                 options=options,
-                value=[value],
+                value=value if isinstance(value, list) else [value],
             )
         else:
             raise ValueError("Unknown dcc component")