+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Fixed
 - links to imprint and privacy page
+- timeseries refresh button
 
 ## [0.14.0] - 2021-10-04
 ### Added